Cleaning up MOAI warnings.

+++ src/moaicore/MOAIProp2D.cpp
@@ -475,7 +475,7 @@ void MOAIProp2D::Draw ( int subPrimID, bool reload ) {
 	
 	if ( this->mGrid ) {
 	
-		if ( subPrimID == MOAIProp::NO_SUBPRIM_ID ) {
+		if ( (u32)subPrimID == MOAIProp::NO_SUBPRIM_ID ) {
 	
 			MOAICellCoord c0;
 			MOAICellCoord c1;
